feat(TodoBox): separate state management logics and component
